requirements.txt
asciimatics==1.14.0
future==0.18.2
numpy==1.23.4
opencv-python==4.6.0.66
Pillow==9.2.0
pyfiglet==0.8.post1
wcwidth==0.2.5
